feat(cloudflare): migrate application data and workflows to D1
Replace PostgreSQL runtime queries and Vercel workflows with native D1 batches, guarded writes, Cloudflare Workflows, and scheduled Worker entrypoints. Preserve billing data, search behavior, dates, and authenticated feature contracts.

Use persisted data caching to avoid workerd stream hangs, bound vector scoring, and verify native database behavior plus authenticated Worker browser flows.

+++ lib/dashboard/intelligence-facebook-pages.ts
@@ -19,7 +19,7 @@ export async function listIntelligenceFacebookPages(): Promise<
 	const [evidenceRows, trackedRows, profiles] = await Promise.all([
 		adminDb
 			.select({
-				evidenceCount: sql<number>`count(*)::int`,
+				evidenceCount: sql<number>`count(*)`,
 				facebookId: facebookIdExpr,
 				lastSeenAt: sql<Date | null>`max(${evidenceItems.createdAt})`,
 				username: evidenceItems.author,
